% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axis_dim.R
\name{axis_dim}
\alias{axis_dim}
\title{create axis}
\usage{
axis_dim(x, chart_type, axis)
}
\description{
create axis
}
\examples{
x <- list(min = -2, max = 6, step = 2, pivot = -2)
axis_dim(x, chart_type = "RowChart", axis = "x")
axis_dim(x, chart_type = "StackedBarChart", axis = "y")
}
