% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Function for creating chart configurations.}
\source{
\code{jsonlite} was created by Jeroen Ooms et al., see
  \url{https://cran.r-project.org/web/packages/jsonlite/index.html}.
}
\usage{
create_config(sdmx_data_query = stop("'sdmx_data_query' must be specified"),
  data_api_endpoint = "http://stats.oecd.org:80/SDMX-JSON/data",
  title = NULL, subtitle = NULL, unit = NULL, source_url = NULL,
  source_label = NULL, logo = TRUE, owner = TRUE,
  type = stop("'type' must be specified"), width = NULL, height = NULL,
  path = NULL)
}
\arguments{
\item{sdmx_data_query}{complete SDMX URL created using
\code{create_query_url}}

\item{title}{character string main title of chart}

\item{subtitle}{character vector chart subtitle}

\item{unit}{character string to specify unit of measure (should be automatic)}

\item{source_url}{character string to add URL below chart}

\item{source_label}{character string to show instead of URL below chart}

\item{logo}{logical include logo}

\item{owner}{logical include copyright information}

\item{type}{one of \code{BarChart}, \code{RowChart}, \code{ScatterChart},
\code{HSymbolChart}, \code{VSymbolChart}, \code{TimelineChart},
\code{StackedBarChart}}

\item{width}{integer chart width in pixel}

\item{height}{integer chart height in pixel}

\item{path}{path to a JSON template that can be read using
\code{jsonlite::fromJSON}}
}
\description{
\code{create_config} creates configuration lists for use with
}
\examples{
query <- "KEI/PRINTO01+PRMNTO01.AUS+AUT+BEL+CAN+CHL+CZE+DNK+EST+FIN+FRA.GP.A/all?startTime=2015&endTime=2015"
chartconfig <- create_config(sdmx_data_query = query, type = "BarChart")
str(chartconfig)

## using custom template
jsonfile <- system.file("templates/default.json", package = "rcw")
chartconfig2 <- create_config(sdmx_data_query = query, type = "BarChart", path = jsonfile)
str(chartconfig2)

}
