% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focus_dim.R
\name{focus_dim}
\alias{focus_dim}
\title{n dimensions}
\usage{
focus_dim(x, chart_type)
}
\arguments{
\item{x}{a list of character vectors respecting the order of dimensions of
the SDMX data flow}
}
\description{
n dimensions
}
\examples{

focus_dim(x = list("PRINTO01"))
focus_dim(x = list("PRINTO01", c("BEL", "FIN")))
focus_dim(x = list("PRINTO01", c("BEL", "FIN"), c("2010")))
focus_dim(x = list("PRINTO01", c("BEL", "FIN"), c("2010"), c("AUT", "AUS", "USA", "CAN"), c("PCT", "ABS")))

## StackedBarChart: only uses first list element

focus_dim(x = list(c("PRINTO01", "PRMNTO01"), c("BEL", "FIN")), chart_type = "StackedBarChart")
focus_dim(x = list(c("PRINTO01", "PRMNTO01"), c("BEL", "FIN")), chart_type = "ScatterChart")
}
