% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/focus_dim.R
\name{focus_dim2}
\alias{focus_dim2}
\title{2 dimensions}
\usage{
focus_dim2(x)
}
\arguments{
\item{x}{a list of character vectors respecting the order of dimensions of
the SDMX data flow}
}
\description{
2 dimensions
}
\examples{

focus_dim2(x = list("PRINTO01", c("BEL", "FIN")))

list(list(value = "PRINTO01 - BEL"),
     list(value = "PRINTO01 - FIN"))
}
