% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_config.R
\name{post_config}
\alias{post_config}
\title{Function for posting chart configurations.}
\source{
\code{httr} was created by Hadley Wickham, see
  \url{https://cran.r-project.org/web/packages/httr/index.html};
  \code{jsonlite} was created by Jeroen Ooms et al., see
  \url{https://cran.r-project.org/web/packages/jsonlite/index.html}.
}
\usage{
post_config(config = stop("'config' must be specified"), ...,
  url = "http://stats.oecd.org", show = FALSE, path = NULL,
  response = FALSE, auto_unbox = TRUE, null = "null")
}
\arguments{
\item{config}{a configuration list according to the API definition}

\item{...}{additional parameters supplied to \code{httr:POST}, e.g.
\code{httr::verbose()}}

\item{url}{the POST API endpoint offering \code{/share} and \code{/chart}
services, e.g. http://vs-webdev-1:89 or http://stats.oecd.org}

\item{show}{boolean print the created JSON object}

\item{path}{write the JSON object to a file on disk}

\item{auto_unbox}{automatically ‘unbox’ all atomic vectors of length 1 (see
\code{jsonlite::toJSON})}

\item{null}{how to encode NULL values within a list: must be one of 'null' or
'list' (see \code{jsonlite::toJSON})}
}
\description{
\code{post_config} submits JSON configuration lists.
}
\examples{
query <- "KEI/PRINTO01+PRMNTO01.AUS+AUT+BEL+CAN+CHL+CZE+DNK+EST+FIN+FRA.GP.A/all?startTime=2015&endTime=2015"
chartconfig <- create_config(sdmx_data_query = query, type = "BarChart", logo = FALSE)
post_config(config = chartconfig, url = "http://stats.oecd.org", response = TRUE)

## write to disk for debugging
outfile <- tempfile(pattern = "chartconfig_", fileext = ".json")
post_config(config = chartconfig, path = outfile)
filecon <- file(outfile)
readLines(con = filecon)
close(filecon)


}
