% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcwChart.R
\name{rcwChart}
\alias{rcwChart}
\alias{rcwChartOutput}
\alias{renderrcwChart}
\title{Function for creating rcw D3 graphs.}
\source{
D3.js was created by Michael Bostock, see \url{http://d3js.org/}.
}
\usage{
rcwChart(data, type = stop("'type' must be provided"), height = "300",
  width = "600")

rcwChartOutput(outputId, width = "100\%", height = "500px")

renderrcwChart(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{data}{a data frame object with three columns. The first column
corresponds to the x-axis, the second to the y-axis and the third indicates
the highlighting pattern: negative integers indicate non-highlighted
values, odd and even numbers are distinguished with permanent color
highlighting, odd numbers include zero.}

\item{type}{a valid chart type e.g. \code{"bar"}, \code{"scatter"},
\code{"row"} etc.}

\item{height}{height for the graph's frame area in pixels (if
\code{NULL} then height is automatically determined based on context)}

\item{width}{numeric width for the graph's frame area in pixels (if
\code{NULL} then width is automatically determined based on context)}
}
\description{
\code{rcwChart} creates simple D3 JavaScript graphs.
}
\examples{
rcwData <- data.frame(
  x = c(1000, 1001, 1010, 1011, 1100, 1101, 1110, 1111),
  y = c(15, -20, 10, 40, 30, 25, 65, 27),
  highlightIndex = c(-1, -1, 0, -1, -1, -1, 1, -1))
rcwChart(rcwData, type = "bar")

}
