% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_config.R
\name{create_config}
\alias{create_config}
\title{Function for creating chart configurations.}
\source{
\code{jsonlite} was created by Jeroen Ooms et al., see
  \url{https://cran.r-project.org/web/packages/jsonlite/index.html}.
}
\usage{
create_config(source = stop("'source' must be specified"), title = NULL,
  subtitle = NULL, unit = NULL, footurl = NULL, footlabel = NULL,
  logo = TRUE, owner = TRUE, type = stop("'type' must be specified"),
  width = NULL, height = NULL, path = NULL)
}
\arguments{
\item{source}{complete SDMX URL created using \code{create_query_url}}

\item{title}{character string main title of chart}

\item{subtitle}{character vector chart subtitle}

\item{unit}{character string to specify unit of measure (should be automatic)}

\item{footurl}{character string to add URL below chart}

\item{footlabel}{character string to show instead of URL below chart}

\item{logo}{logical include logo}

\item{owner}{logical include copyright information}

\item{type}{one of \code{BarChart}, \code{RowChart}, \code{ScatterChart},
\code{HSymbolChart}, \code{VSymbolChart}, \code{TimelineChart},
\code{StackedBarChart}}

\item{width}{integer chart width in pixel}

\item{height}{integer chart height in pixel}

\item{path}{path to a JSON template that can be read using
\code{jsonlite::fromJSON}}
}
\description{
\code{create_config} creates configuration lists for use with
}
\examples{
flow <- "KEI"
query <- "PRINTO01+PRMNTO01.AUS+AUT+BEL+CAN+CHL+CZE+DNK+EST+FIN+FRA.GP.A"
opts <- c("startTime=2015", "endTime=2015", "dimensionAtObservation=AllDimensions")
query_url <- create_query_url(flow = flow, query = query, opts = opts)
chartconfig <- create_config(source = query_url, type = "BarChart")
str(chartconfig)

## using custom template
jsonfile <- system.file("templates/default.json", package = "rcw")
chartconfig2 <- create_config(source = query_url, type = "BarChart", path = jsonfile)
str(chartconfig2)

}
