% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_query_url.R
\name{create_query_url}
\alias{create_query_url}
\title{Function for creating SDMX queries.}
\usage{

  create_query_url(sdmx_data_query = stop("'sdmx_data_query' must be provided"),
  data_api_endpoint = stop("'data_api_endpoint' must be provided"))
}
\arguments{
\item{data_api_endpoint}{SDMX API URL, e.g. \url{http://stats.oecd.org:80/SDMX-JSON/data}}

\item{query}{SDMX query, e.g.
\code{"KEI/PRINTO01+PRMNTO01.AUS+AUT+BEL+CAN+CHL+CZE+DNK+EST+FIN+FRA.GP.A/all?startTime=2015&endTime=2015"}}
}
\description{
\code{create_query} creates SDMX query URLs
}
\examples{
query <- "KEI/PRINTO01+PRMNTO01.AUS+AUT+BEL+CAN+CHL+CZE+DNK+EST+FIN+FRA.GP.A/all?startTime=2015&endTime=2015"
create_query_url(sdmx_data_query = query,
                 data_api_endpoint = "http://stats.oecd.org:80/SDMX-JSON/data")

}
